Support reduced mode without cancel or feedback
The many topic connections for actionlib can be a problem. We can now omit
creation of a cancel publisher or feedback subscriber in the client, to
reduce that number. We have seen that they are not always required, i.e.
for actions that do not provide intermediate feedback and that cannot be
cancelled. Both holds true for the converted appletlib actions.